# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hensun Desk Cloud",
    "description": "Commercial pilot control plane; not a production certification.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/memories/export": {
      "get": {
        "tags": [
          "memory-portability"
        ],
        "summary": "Export Memories",
        "operationId": "export_memories_v1_memories_export_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryExportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/memories": {
      "delete": {
        "tags": [
          "memory-portability"
        ],
        "summary": "Delete All Account Memories",
        "operationId": "delete_all_account_memories_v1_memories_delete",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/conversations/{conversation_id}/summary": {
      "put": {
        "tags": [
          "memory-portability"
        ],
        "summary": "Update Session Summary",
        "operationId": "update_session_summary_v1_conversations__conversation_id__summary_put",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionSummaryUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Update Session Summary V1 Conversations  Conversation Id  Summary Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "memory-portability"
        ],
        "summary": "Delete Session Summary",
        "operationId": "delete_session_summary_v1_conversations__conversation_id__summary_delete",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/ota/releases": {
      "post": {
        "tags": [
          "ota"
        ],
        "summary": "Register Release",
        "operationId": "register_release_v1_ota_releases_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirmwareReleaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirmwareReleaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ota/check": {
      "get": {
        "tags": [
          "ota"
        ],
        "summary": "Check Release",
        "operationId": "check_release_v1_ota_check_get",
        "parameters": [
          {
            "name": "current_version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Current Version"
            }
          },
          {
            "name": "Device-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/FirmwareReleaseResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Check Release V1 Ota Check Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FirmwareReleaseRequest": {
        "properties": {
          "board_type": {
            "type": "string",
            "pattern": "^[a-z0-9.-]+$",
            "title": "Board Type"
          },
          "version": {
            "type": "string",
            "pattern": "^[0-9]+\\.[0-9]+\\.[0-9]+(?:[-+][a-zA-Z0-9.-]+)?$",
            "title": "Version"
          },
          "artifact_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Artifact Url"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Sha256"
          },
          "signature": {
            "type": "string",
            "minLength": 32,
            "title": "Signature"
          },
          "rollout_percent": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Rollout Percent"
          },
          "mandatory": {
            "type": "boolean",
            "title": "Mandatory",
            "default": false
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": false
          },
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "board_type",
          "version",
          "artifact_url",
          "sha256",
          "signature",
          "rollout_percent"
        ],
        "title": "FirmwareReleaseRequest"
      },
      "FirmwareReleaseResponse": {
        "properties": {
          "board_type": {
            "type": "string",
            "title": "Board Type"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "artifact_url": {
            "type": "string",
            "title": "Artifact Url"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          },
          "rollout_percent": {
            "type": "integer",
            "title": "Rollout Percent"
          },
          "mandatory": {
            "type": "boolean",
            "title": "Mandatory"
          }
        },
        "type": "object",
        "required": [
          "board_type",
          "version",
          "artifact_url",
          "sha256",
          "signature",
          "rollout_percent",
          "mandatory"
        ],
        "title": "FirmwareReleaseResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoryExportResponse": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "agent_memories": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Agent Memories"
          },
          "session_summaries": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Session Summaries"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "agent_memories",
          "session_summaries"
        ],
        "title": "MemoryExportResponse"
      },
...
      "SessionSummaryUpdateRequest": {
        "properties": {
          "summary": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "SessionSummaryUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}